Reuse shared file type mapping in tool blocks
